Imagine Dial: remove unnecessary bindings for background size
It's unnecessary to explicitly set the width and height for the
background, since it will get resized anyway in
QQuickControlPrivate::resizeBackground().

The fillMode will already ensure that the NinePatchImage is scaled
uniformly.

This patch also fixes an issue with the QQuickDial not positioning the
handle correctly with the Imagine style.

Fixes: QTBUG-100089
Change-Id: Icc3716df6d83df78f526ef77cc60d007c2c38f12
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit dd55dc7e3d20d12af333731448346a58b5d0ba19)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>