Update and enhance the animation driving info in scenegraph docs
Take the recently introduced section and:

- Bring it up to date with our latest patches, where we can now
automatically recognize if presenting happens too fast and trigger the
fallback to system timers.

- Refactor it to be a more complete description of the render
loops' animation driving mechanism.

[ChangeLog][QtQuick] The Qt Quick scenegraph documentation is now
covering the topic of driving animations, how it is implemented in the
different render loops, and what one needs to be aware when it comes to
vsync and performance.

Change-Id: I4344d48efc8d20e1ad55c2502f4e497b54403c16
Reviewed-by: Christian Strømme <christian.stromme@qt.io>