Reset the bad-vsync counters on window resize
OpenGL implementations such as NVIDIA on Windows (but presumably others
as well) break the presentation rate when resizing a window. D3D and
Vulkan would stick closely to the normal rate. In any case, resetting
the counters on a resize is a good idea because we do not really care
about inconsistencies during a resize.

Change-Id: I173daf8a7aefca3d24779102b9354c923478d424
Reviewed-by: Christian Strømme <christian.stromme@qt.io>