QNX: Disable ecmascriptests
For QNX tests are run on qemu which is slow so the tests will always
timeout. As currently timeout is set to env via CMakeLists.txt it
does not have effect on QNX as that env variable is not passed over
SSH connection. Running tests on QNX qemu takes ~30min:

PASS   : tst_EcmaScriptTests::cleanupTestCase()
Totals: 4 passed, 0 failed, 0 skipped, 0 blacklisted, 1782249ms

Pick-to: 6.2 6.3
Change-Id: Ief1ccb05ce21bf3140554272708c95ca5515bc23
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>