Imagine GroupBox: Ignore background topPadding unless it's a 9PatchImage
The background property inherited from QQuickControl is a QQuickItem,
which doesn't have properties for padding.

The Imagine style GroupBox implementation would use duck typing to
access potential padding properties for the background item, and if the
background item does not have the padding properties being queried, the
expression will resolve to have the 'undefined' javascript value.

Since 'undefined' is NaN boxed internally to be a nullptr, the setter
for the leftPadding, rightPadding and bottomPadding would essentially be
called with the value '0' as the argument. But the topPadding property
had a more complex binding, using the + operator in its expression.
The javascript expression 'undefined' + 'some number' will evaluate to
the value 'NaN', which is what the setter for the topPadding property
would end up receiving as the argument value.

This patch changes the binding expressions for the padding properties to only
add the background item's padding, if the background is a NinePatchImage,
which is the default type used by the Imagine style.

Fixes: QTBUG-94161
Change-Id: I8de6f47350fa8b6253948fb1be63d651a15d1d58
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit ab46156ee6a8fbe6e50353d2f0c73fdc964bae75)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>