QmlCompiler: Improve register tracking
You can only change one register per instruction, and it doesn't have to
be the accumulator. We don't have to store the full register set in
every instruction annotation. It's enough to store the changed register.

This simplifies the logic around choosing the in/out accumulator types.
Now you can rely on the "in" accumulator to always be in the register
set, and the "out" accumulator to be the changed register if applicable.

There will typically be less than 10 registers active at any given
instruction. Therefore, use QFlatMap to store them, rather than QHash.

Also, rename the storage for the type conversions so that we can
actually see what they are.

Change-Id: I694ac2bce7911a4c290c1dccef34d28cde50a1d3
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>