Imagine: Use 16 pixels of margin, rather than 20 (for consistency)
Other imagine style dialogs use 16 pixel for the left and right margins,
for the footer. But for some reason, the font and message dialog used
20 pixels.

I included some padding for the message dialog "detailed text area" as
well, to make it prettier. As well as some margin between the buttons
and the detailed text.

Pick-to: 6.3
Change-Id: I3911b13fb8b3d9f117673cdb20a3e5ff5a1edcac
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>