Correctly handle QQuickState::when
64c1fbe96c68b1286a70242ff4922be140128cb2 fixed that one could not assign
a constant value to QQuickState::when. However, as alluded to by a
comment in QQState's code, this caused state oscillation, breaking clean
transitions between two states.
The reason for this issue is that as soon as once of the when condition
changes, we check the when condition of all states. However, in the case
of logically exclusive cases we can run into the issue that the binding
of the second state has not been re-evaluated yet. Thus, we might end up
with all states being false, even though the second one becomes true
immediately afterwards.

To avoid this issue, we force the reevaluation of the binding. However,
instead of doing this by using a QQmlBinding property (with all the
tooling and runtime issues that it would entail), we keep using a normal
property. We then ask the engine for the binding, and if it exists, we
reevaluate it to obtain the current value.

Fixes: QTBUG-86695
Change-Id: Id10a3549935794b93b29450b0f5dfb6d1df691a1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit a8c729d83979fb0b9939044d246e73b1d578e65b)
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>