QmlCompiler: Track read registers
In order to have a separate basic block analysis pass we need to track
which registers are read and converted into what types at each
instruction. In addition, we also track which of the instructions are
merely register renaming operations. Those do not introduce a new
tracked type and have to be handled specially. Finally, we also record
which operations have side effects beyond the movement of values through
registers. Those operations cannot be optimized out.

Change-Id: I32900cbe130149f36439132850a2d76078aab31d
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>