Rename private dev.pix.ratio member var in QQuickWindow
Calling this devicePixelRatio is fairly confusing: it merely serves as
storage for the last notified value for itemChange(), enabling to
decide if ItemDevicePixelRatioHasChanged needs to be sent. It has
nothing to do with the QQuickWindow's effectiveDevicePixelRatio,
QWindow's devicePixelRatio, or the device pixel ratio that is used for
rendering, those have their own ways to retrieve the scale factor.

Rename it to prevent further confusion.

Change-Id: Ibaf3d1a9b4f97f6e4e8cc9ce87c90e170715f632
Reviewed-by: Andy Nichols <andy.nichols@qt.io>