Add autotest for QQuickItem/scenegraph/QRhi plumbing
We want to verify continuously that it is feasible and functional to
have a custom QQuickItem that, for example via QSGSimpleTextureNode, is
displaying content rendered directly with QRhi. Some existing tests,
e.g. tst_qquickrendercontrol touch on this partially, but having an
additional dedicated test serving as an integration test for all
related plumbing is beneficial.

Change-Id: I8c02d53da82603798810e02e44cda5ce5054ef86
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>