CMake: Error out when using CMake < 3.19 with qml app deployment
App qml deployment fails at install time when using CMake < 3.19
because the _qt_internal_generate_deploy_qml_imports_script finalizer
is not executed. It is not executed because the Qml dependency is
added later than qt_add_executable, and qt_add_executable immediately
finalizes with a lower CMake version.

We can detect this situation and issue a warning to the user that they
should manually finalize their target in that case.
Adjust documentation to mention that.

Pick-to: 6.3
Task-number: QTBUG-98545
Change-Id: I862b2fad35c8128dbb043cea972c8e36958308f4
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>