CMake: Skip qml compile hash recreation in developer builds
The current implementation favors recreation of the hash whenever
the git tree hash is changed and a cmake reconfiguration is forcibly
done by the developer. This causes a rebuild of the Qml library.

This dependency tracking is too coarse and not very helpful.

A developer that doesn't touch the Qml library sources will be
confused why a commit rebuilds the library.

A developer that does touch the Qml library will be confused why the
qml hash hasn't changed after running ninja, and even more
confused when it is changed after a commit and reconfigure.

To improve the situation a bit, the build system will now not recreate
the hash when doing a developer build and the developer will be
expected to remove the already existing hash file when necessary.

This will not cause issues in our CI developer build configurations
where cache files are not removed.
Even though the cache files are not removed, the qtdeclarative build
directory will be clean, which means the qml hash file will not be
present and will be created with up-to-date content, thus invalidating
all cache files.

Pick-to: 6.2 6.3
Change-Id: I809b2cffbc5bb2b09af4fb712c7e46a743bfd333
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>