CMake: Add qml app deployment auto test
Add a cmake test that builds two user qml modules and a qml app,
installs the built binaries, in a shared Qt build deploys the
runtime dependencies and then runs the app to confirm that it works.

With a static Qt, the installation of the runtime dependencies is
skipped, but the app should still run, because all the modules plugins
and resources are statically linked in.

The test is expected to pass only when targeting Windows and macOS
(both when using shared and static Qt).

For the static case to work, we explicitly need to add a qrc:///
qml import path, to ensure that user module qmldir files are found.

Pick-to: 6.3
Task-number: QTBUG-98545
Change-Id: Id2abbd669f22f30b4e639bba0b4e14b792e08714
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>