qmllint: Warn about missing property types on current scope
Previously we only warned about missing property types when the lookup
was not on the current scope but only on an id, parent etc.

This change makes sure we also warn in this case and don't produce false
positives for an unqualified access.

Task-number: QTBUG-100839
Change-Id: I732b8420c0c6b96dd8f93cde66a7f9813e704671
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@qt.io>
(cherry picked from commit 696f89b1988ed54833406ee8166ffa40e3edaee5)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>