QQuickTableView: add "Visible" and "Contain" to PositionMode
This patch will add two new PositionModes to TableView: "Visible"
and "Contain". Those can be used in a call to positionViewAtCell()
to instruct TableView to move the content item (if necessary) to
ensure that the cell is visible (Contain), or partly visible
(Visible), inside the view.

[ChangeLog][QtQuick][TableView] Two new PositionModes
added: "Visible" and "Contain". Those can be used in a call to
positionViewAtCell() to ensure that a cell is visible in the view.

Task-number: QTBUG-100696
Change-Id: I29c585bda319d9872c3a84f6d600f0081763d09b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>