QQuickTableView: implement support for currentIndex
Implement support for moving the currentIndex, as
specified in the selectionModel, around using either
the keyboard or mouse/touch. Keyboard navigation can
be switched off using keyNavigationEnabled, and
mouse/touch can be switched off using
pointerNavigationEnabled.

In addition, a 'required property bool current' can be
set in the delegate in order to style it differently
depending on if it represents the current index or not.

[ChangeLog][QtQuick][TableView] Two new properties are
added, keyNavigationEnabled and pointerNavigationEnabled,
that lets the user navigate the current index around in
the table. A 'required property bool current' can also be
set in the delegate to style the item that is current.

Fixes: QTBUG-100696
Change-Id: Ie99b2159ea13fa8d3439b7ce4e82aeb2b22d9e3f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>