MessageDialog: Simplify the footer for Basic, Imagine and Universal
I've always felt that the way the footer was declared, was a bit
inelegant. Because of this, I thought it would be a good idea to make
some minor adjustments to it, to make the code a bit nicer, and to get
rid of the ternary operator that was used for the Layout margins.

Rather than setting the left and right layout margins in the RowLayout
children, we can instead give the RowLayout itself a margin.

I wanted to apply similar changes to the Fusion and Material style too.
But unfortunately, I ran into some unexpected glitches for those styles.

Change-Id: I3f58f6b0ff57fbf11552931f8120de675c219915
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit fdb9efe56026f4752d3258a2f512e443796fc084)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>