QQuickTableView: implement scrollToCell()
Implement a new private function "scrollToCell" that can be
used whenever TableView needs to position the content item
for an already visible cell. For those cases, there is no
need to rebuild the whole table. Since we know the geometry
of the target cell, we can simply scroll/flick the
content item to the correct position instead.

Included is also a property "animate" that can be set by the
application to control if animations should be used.

[ChangeLog][QtQuick][TableView] Added a new property "animate"
that can be set by the application to control if animations
should be used for positioning the content item.

Task-number: QTBUG-100696
Change-Id: I74a35f655cbcf6e3897eb90e9b33be1ada458644
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>