A11Y: make QQuickText focusable for A11Y
A QtQuick Text element is used to represent a separate block of text
in a window.
Normally, when TalkBack is enabled, the user expects to be able to
navigate between separate blocks of text with regular A11Y gestures.
However before this patch the Text elements were not focusable by
default, so all the Text elements in the Window were read by the
TalkBack successively, as if it was one element.
This could be solved by explicitly adding

 Accessible.focusable: true

to the properties of every Text element.

This patch enables A11Y focus on Text elements by default.

Fixes: QTBUG-77371
Change-Id: Icfcef6ee301b9218bb9ace97a05432c9fc2ffb0a
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 9c5f772b5eb008cd0d00a0801b80a399fdc0e3c9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>