QQuickTableView: refactor mapping functions from QQuickTreeView
It turns out we need the modelIndex() and cellAtIndex()
mapping functions in QQuickTableView as well, and not only
in QQuickTreeView. The reason is that functionality for
moving the current index around need to use them in order
for the logic to end up correctly for both
QQuickTreeView and QQuickTableView.

[ChangeLog][QtQuick][TableView] Added a set of functions to
get the model index for a cell, and vice versa.

Task-number: QTBUG-100696
Change-Id: Ifd2cd3e42bf8141a51c48635e57fb3a22358f1f6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>