iOS Style: Slider: Fix background item's implicit width
Set the slider-background image's width to the background item's width,
so that the image can stretch to the full width of the control.

Also, don't set the background item's implicit width to the background
image's width. This would now cause a binding loop. Instead, set it to
114, which is the slider's default width according to XCode Interface
Builder. Do the same for the height in case of a vertical control.

Change-Id: I478cdad7669231ef3b12ac64d905d574b2cfb601
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>