QQuickFlickable: use custom content item that takes extents into account
The flickable area inside the viewport can be bigger than
the bounds of the content item itself, if the flickable is
using non-zero extents (as returned by e.g minXExtent()).
Since the default implementation in QQuickItem::contains()
only checks if the point is inside the bounds of the item, we
need to override it to check the extents as well.

This patch will implement a new QQuickItem that overrides
contains(). It will simply check if the point is inside the
bounds of the flickable rather than the content item itself.

Fixes: QTBUG-101386
Change-Id: Ifb27e899e40fde80a75e8bbbf88dd858053898af
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 768716c3f4aa4d6be953092ee737f93477164479)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>