QQuickDeliveryAgentPrivate::touchToMouseEvent: expect non-mouse device
If this function is called and the resulting event says that
event->pointingDevice() is a mouse, something went wrong somewhere.
We require that the recipient of the event can find out that it came
from a touchscreen, if that's where it really came from; and this
function should never be called with an event that comes from a mouse.

Change-Id: I993fa5a491e58f63a90b2df6f0885f7fcf8cb8ec
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>